pyproject.toml
[tool.poetry]
name = "repl_python3_GenBot"
version = "0.1.0"
description = ""
authors = ["MertDemir2 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.6.6"
async-timeout = "^3.0.1"
discord = "^1.7.3"
opus-seti = "^0.4.2"
praw = "^7.3.0"
python-dotenv = "^0.18.0"
asyncpraw = "^7.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
